Change brightness/contrast range from -1.0..1.0 to -100..100
Updated brightness and contrast to use integer values from -100 to +100
instead of float values from -1.0 to 1.0, matching ImageMagick's native
format and providing finer control for users.

Rust changes:
- Changed CLI parameters from f32 to i32 (-100 to +100)
- Updated ProcessingConfig to use i32 for brightness/contrast
- Modified apply_brightness_adjustment() and apply_contrast_adjustment()
  to accept i32 directly (no conversion needed for ImageMagick)
- Updated fallback implementations to convert i32 to appropriate ranges
- Fixed all type conversions in processing pipeline
- Auto-optimizer results still use f32, converted to i32 when applied

Flutter changes:
- Changed ProcessingConfig fields from double to int
- Updated sliders to range 0-200 (mapping to -100..+100)
- Display shows integer values directly (no decimal places)
- Regenerated JSON serialization code

Benefits:
- Native ImageMagick format (no conversion overhead)
- Finer granular control (201 steps vs 201 float precision)
- More intuitive for users (whole numbers instead of decimals)
- Matches professional image editing tools

+++ photoframe_flutter/lib/models/processing_config.dart
@@ -40,8 +40,8 @@ class ProcessingConfig {
   final bool autoColorCorrect;
   final DitherMethod ditherMethod;
   final double ditherStrength;
-  final double contrast;
-  final double brightness;
+  final int contrast;
+  final int brightness;
   final bool autoOptimize;
   final bool outputBmp;
   final bool outputBin;
@@ -73,8 +73,8 @@ class ProcessingConfig {
     this.autoColorCorrect = true,
     this.ditherMethod = DitherMethod.jarvisJudiceNinke,
     this.ditherStrength = 1.0,
-    this.contrast = 0.0,
-    this.brightness = 0.0,
+    this.contrast = 0,
+    this.brightness = 0,
     this.autoOptimize = false,
     this.outputBmp = false,
     this.outputBin = true,
@@ -112,8 +112,8 @@ class ProcessingConfig {
     bool? autoColorCorrect,
     DitherMethod? ditherMethod,
     double? ditherStrength,
-    double? contrast,
-    double? brightness,
+    int? contrast,
+    int? brightness,
     bool? autoOptimize,
     bool? outputBmp,
     bool? outputBin,
